storyweaver/requirements.txt
gradio>=4.44.0
requests>=2.31.0
python-dotenv>=1.0.1
